feat: streams::document_event
