+++ src/app/streams.rs
@@ -1,5 +1,11 @@
+use crate::browser::util::{document, window};
+use crate::virtual_dom::Ev;
 use futures::stream::{Stream, StreamExt};
 use gloo_timers::future::IntervalStream;
+use web_sys::Event;
+
+mod event_stream;
+use event_stream::EventStream;
 
 // ------ Interval stream ------
 
@@ -27,5 +33,48 @@ pub fn interval<MsU>(
 
 // ------ Window Event stream ------
 
-mod window_event;
-pub use window_event::window_event;
+/// Stream `Window` `web_sys::Event`s.
+///
+/// Handler has to return `Msg`, `Option<Msg>` or `()`.
+///
+/// # Example
+///
+/// ```rust,no_run
+///orders.stream(streams::window_event(Ev::Resize, |_| Msg::OnResize));
+///orders.stream_with_handle(streams::window_event(Ev::Click, |_| log!("Clicked!")));
+/// ```
+///
+/// # Panics
+///
+/// Panics when the handler doesn't return `Msg`, `Option<Msg>` or `()`.
+/// (It will be changed to a compile-time error).
+pub fn window_event<MsU>(
+    trigger: impl Into<Ev>,
+    handler: impl FnOnce(Event) -> MsU + Clone + 'static,
+) -> impl Stream<Item = MsU> {
+    EventStream::new(&window(), trigger.into()).map(move |event| handler.clone()(event))
+}
+
+// ------ Document Event stream ------
+
+/// Stream `Document` `web_sys::Event`s.
+///
+/// Handler has to return `Msg`, `Option<Msg>` or `()`.
+///
+/// # Example
+///
+/// ```rust,no_run
+///orders.stream(streams::document_event(Ev::SelectionChange, |_| Msg::OnSelection));
+///orders.stream_with_handle(streams::document_event(Ev::SelectionChange, |_| log!("Selection changed!")));
+/// ```
+///
+/// # Panics
+///
+/// Panics when the handler doesn't return `Msg`, `Option<Msg>` or `()`.
+/// (It will be changed to a compile-time error).
+pub fn document_event<MsU>(
+    trigger: impl Into<Ev>,
+    handler: impl FnOnce(Event) -> MsU + Clone + 'static,
+) -> impl Stream<Item = MsU> {
+    EventStream::new(&document(), trigger.into()).map(move |event| handler.clone()(event))
+}
